refactor: take reviews
